requirements.txt
# Core dependencies (no torch needed)
faiss-cpu==1.12.0
llama-index==0.11.13
llama-index-core==0.11.13.post1
llama-index-vector-stores-faiss==0.2.1
llama-index-readers-file==0.2.2
openai==1.98.0
pandas==2.3.1
orjson==3.10.7
tiktoken==0.11.0
tqdm==4.67.1

# Streamlit for web interface
streamlit==1.39.0